feat(wallet): ser send_transaction
